feat: enhance sandbox preview and proxy functionality
- Added comprehensive tests for creating and managing sandbox previews, including handling undeclared ports and latency comparisons.
- Introduced new tests for proxy configurations, validating header and body injections, and ensuring proper routing through proxies.
- Implemented end-to-end tests for proxy functionality with various CLI tools and agents, ensuring correct behavior and performance metrics.
- Removed outdated proxy tests and consolidated proxy-related tests into dedicated files for better organization and clarity.
